chore: Kyle PR review comments
